#96 fix: 연차 중복체크 Jpa 오류 수정

+++ src/main/java/com/example/miniproject/annual/repository/AnnualRepository.java
@@ -18,5 +18,5 @@ public interface AnnualRepository extends JpaRepository<Annual, Long> {
 
 	@Query("SELECT count(a) FROM Annual a "
 		+ "  WHERE ((a.startedAt < :endDate AND a.lastedAt > :startDate) OR (a.startedAt = :startDate and a.lastedAt = :endDate)) AND a.member = :member")
-	int countByStartedAtAndLastedAt(LocalDate startDate, LocalDate endDate, Member member);
+	Long countByStartedAtAndLastedAt(LocalDate startDate, LocalDate endDate, Member member);
 }
